Add Support for REST API
-Add Spring boot support
-Add Service and Controller class

+++ src/main/java/com/codecatalyst/net/NetUtils.java
@@ -5,7 +5,7 @@ public class NetUtils {
         String[] parts = ip.split("\\.");
         long result = 0;
         for (int i = 0; i < parts.length; i++) {
-            result += Integer.parseInt(parts[i]) * Math.pow(256, 3 - i);
+            result += (long) (Integer.parseInt(parts[i]) * Math.pow(256, 3 - i));
         }
         return result;
     }
